docs: labs: add a pointer to config.md
The documentation mentions that Labs features can be manipulated by
opening the Settings menu and selecting Labs.  This feature is normally
disabled, so for most users this is not the case.

Add a link to the `config.md` file and a note that the Labs features
must be enabled in order to manipulate them.

Signed-off-by: Sean Cross <sean@xobs.io>