Take encrypted message search out of labs
Fixes https://github.com/vector-im/riot-web/issues/13262

This is part of the cross-signing shipping master plan. Known issues relating to this feature are:
* https://github.com/vector-im/riot-web/issues/12896
* https://github.com/vector-im/riot-web/issues/12385
* https://github.com/vector-im/riot-web/issues/11831
* https://github.com/vector-im/riot-web/issues/11155

In theory, these are issues we're comfortable with shipping as we're already enabling it by default. This just makes it easier on everyone by removing the flag (making it still enabled by default).