Demonstrate dis.fire() with view_user_settings
Like ae75cae64cfe29d537706e24231307f2f8d8710d, this proves that the new dispatcher conversion works for fire-and-forget style dispatches too. This is another obvious-if-broken and generally safe conversion to make.

Other actions which can be dispatched this way have been excluded for reasons mentioned in the Action enum's comments.