Fix exceptions from Tooltip
renderTooltip was not a bound function and so was failing to find
the parent when called from the 'scroll' event listener because
'this' was the window object rather than the Tooltip object.

Unsure at what point this broke - I assumed it was in thr recent
typescript conversion but it looks like it would have had the same
problem before.

(cherry picked from commit 0427198f4ae84cad86ec9d18fd28a05dfd4046ad)