Ensure message previews update when needed
In e3a412fbb16321ddd9413a9f3f0c11807c6bd37b we stopped updating the sublist whenever we felt like it, which indirectly froze message previews for room tiles (badges, unread state, etc were unaffected because that is managed by a different store). To fix this, we simply have to listen for changes and perform an update.