Handle tag changes in sticky room updates
Fixes https://github.com/vector-im/element-web/issues/14848

When we're filtering the sticky room will be excluded from the filtered set, and thus won't even appear in the `getOrderedRoomsWithoutSticky()` result. Further, we will likely have to update the position ourselves to ensure the sticky room can be placed appropriately in the list.