Remove some unused components
We no longer have tinting support, so remove it. We still need the `Tinter` to exist though as it's used in quite a few places (though does nothing). Similarly, we have to keep the `roomColor` setting due to it being used in a few places - another PR can take away the tinter support properly.

The room tile context menu and top left menu are artifacts of the old room list.

The end to end tests weren't failing before as the code path is unused, however it seems worthwhile to keep it as we will eventually need it.