Tune cross-signing toasts when 4S is missing
For the case where cross-signing is trusted on device but secret storage does
not exist, we were showing "verify this device", which is not the best match
from the existing toasts. This tunes the checks to instead show "set up
encryption" which is at least a bit closer.

Part of https://github.com/vector-im/element-web/issues/14970